requirements.txt
# Django==1.7.4
https://www.djangoproject.com/download/1.8b1/tarball
django-envsettings==0.1
whitenoise==1.0.6
gunicorn==19.2.1
